# api/openapi.yaml
openapi: 3.1.0
info:
  title: Prompt Engine API
  description: 設計書§13 API設計（REST）
  version: v1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- ciap-bearer: []
paths:
...
  /api/v1/prompts/{namespace}/{name}/versions/{version}/publish:
    post:
      tags:
      - version-controller
      operationId: publish
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeySemVerResponseDto"
...
  /api/v1/prompts/{namespace}/{name}/render:
    post:
      tags:
      - pipeline-controller
      operationId: render
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RenderResponseDto"
  /api/v1/prompts/{namespace}/{name}/execute:
    post:
      tags:
      - pipeline-controller
      operationId: execute
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExecuteRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ExecuteResponseDto"
  /api/v1/prompts/{namespace}/{name}/compile:
    post:
      tags:
      - pipeline-controller
      operationId: compile
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CompileResponseDto"
  /api/v1/prompts/{namespace}/{name}/aliases:
    post:
      tags:
      - alias-controller
      operationId: setAlias
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetAliasRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SetAliasResponseDto"
  /api/v1/prompts/{namespace}/{name}:
    get:
      tags:
      - prompt-controller
      operationId: get
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PromptDetailDto"
    delete:
      tags:
      - prompt-controller
      operationId: archive
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: string
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
    patch:
      tags:
      - prompt-controller
      operationId: updateMetadata
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptMetadataRequestDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/KeyOnlyResponseDto"
...
  /api/v1/prompts/{namespace}/{name}/dependencies:
    get:
      tags:
      - dependency-controller
      operationId: list
      parameters:
      - name: namespace
        in: path
        required: true
        schema:
          type: string
      - name: name
        in: path
        required: true
        schema:
          type: string
      - name: direction
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DependencyEdgeDto"
...
components:
  schemas:
    ContextRequirementDto:
      type: object
      properties:
        scope:
          type: string
        required:
          type: array
          items:
            type: string
        optional:
          type: array
          items:
            type: string
      required:
      - optional
      - required
      - scope
...
    OutputDeclarationDto:
      type: object
      properties:
        format:
          type: string
        schemaRef:
          type: string
      required:
      - format
    ValidationSettingsDto:
      type: object
      properties:
        maxLength:
          type: integer
          format: int32
        maxTokens:
          type: integer
          format: int32
        policies:
          type: array
          items:
            type: string
        placeholders:
          type: string
      required:
      - placeholders
      - policies
    VariableDefinitionDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        source:
          type: string
        required:
          type: boolean
        default:
          type: object
        constraints:
          type: array
          items:
            type: string
        sensitive:
          type: boolean
      required:
      - constraints
      - name
      - required
      - sensitive
      - source
      - type
    KeySemVerResponseDto:
      type: object
      properties:
        key:
          type: string
        semVer:
          type: string
      required:
      - key
      - semVer
...
    OptionsDto:
      type: object
      properties:
        optimize:
          type: boolean
        tokenBudget:
          type: integer
          format: int32
          maximum: 1000000
          minimum: 1
      required:
      - optimize
    RenderRequestDto:
      type: object
      properties:
        versionRef:
          type: string
          minLength: 1
        parameters:
          type: object
          additionalProperties:
            type: object
        context:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
        modelProfile:
          type: string
          minLength: 1
        options:
          $ref: "#/components/schemas/OptionsDto"
      required:
      - context
      - modelProfile
      - parameters
      - versionRef
    FindingDto:
      type: object
      properties:
        rule:
          type: string
        path:
          type: string
        severity:
          type: string
        message:
          type: string
      required:
      - message
      - path
      - rule
      - severity
    RenderResponseDto:
      type: object
      properties:
        promptKey:
          type: string
        version:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/RenderedMessageDto"
        outputFormat:
          type: string
        outputSchemaRef:
          type: string
        tokenEstimate:
          type: integer
          format: int32
        renderHash:
          type: string
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/FindingDto"
        traceId:
          type: string
      required:
      - messages
      - promptKey
      - traceId
      - warnings
    RenderedMessageDto:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
      - content
      - role
    ExecuteRequestDto:
      type: object
      properties:
        versionRef:
          type: string
          minLength: 1
        parameters:
          type: object
          additionalProperties:
            type: object
        context:
          type: object
          additionalProperties:
            type: object
            additionalProperties:
              type: object
        modelProfile:
          type: string
          minLength: 1
        options:
          $ref: "#/components/schemas/OptionsDto"
        executionPolicy:
          $ref: "#/components/schemas/ExecutionPolicyDto"
        outputSchema:
          $ref: "#/components/schemas/OutputSchemaDto"
      required:
      - context
      - executionPolicy
      - modelProfile
      - parameters
      - versionRef
    ExecutionPolicyDto:
      type: object
      properties:
        timeoutMs:
          type: integer
          format: int64
          maximum: 300000
          minimum: 1
        maxRetries:
          type: integer
          format: int32
          maximum: 10
          minimum: 0
        parseRepair:
          type: boolean
      required:
      - parseRepair
      - timeoutMs
    OutputSchemaDto:
      type: object
      properties:
        id:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/OutputSchemaFieldDto"
      required:
      - fields
      - id
    OutputSchemaFieldDto:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        required:
          type: boolean
      required:
      - name
      - required
      - type
    ExecuteResponseDto:
      type: object
      properties:
        promptKey:
          type: string
        version:
          type: string
        outputFormat:
          type: string
        parsedOutput:
          type: object
          additionalProperties:
            type: object
        rawContent:
          type: string
        usage:
          $ref: "#/components/schemas/UsageDto"
        latencyMs:
          type: integer
          format: int64
        evaluationId:
          type: string
        traceId:
          type: string
      required:
      - parsedOutput
      - promptKey
      - rawContent
      - traceId
    UsageDto:
      type: object
      properties:
        inputTokens:
          type: integer
          format: int32
        outputTokens:
          type: integer
          format: int32
        cost:
          type: number
      required:
      - cost
      - inputTokens
      - outputTokens
    CompileResponseDto:
      type: object
      properties:
        promptKey:
          type: string
        traceId:
          type: string
        validationPassed:
          type: boolean
        warningCount:
          type: integer
          format: int32
      required:
      - promptKey
      - traceId
      - validationPassed
      - warningCount
    SetAliasRequestDto:
      type: object
      properties:
        alias:
          type: string
          minLength: 1
        version:
          type: string
          minLength: 1
      required:
      - alias
      - version
    SetAliasResponseDto:
      type: object
      properties:
        key:
          type: string
        alias:
          type: string
        semVer:
          type: string
      required:
      - alias
      - key
      - semVer
    UpdatePromptMetadataRequestDto:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
      - name
      - tags
    KeyOnlyResponseDto:
      type: object
      properties:
        key:
          type: string
      required:
      - key
...
    ExtendsRefDto:
      type: object
      properties:
        key:
          type: string
        range:
          type: string
      required:
      - key
    PromptDetailDto:
      type: object
      properties:
        metadata:
          $ref: "#/components/schemas/PromptMetadataDto"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/PromptVersionDto"
      required:
      - versions
    PromptMetadataDto:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
      required:
      - key
      - name
      - tags
    PromptVersionDto:
      type: object
      properties:
        semVer:
          type: string
        state:
          type: string
        source:
          type: string
        contentHash:
          type: string
        variables:
          type: array
          items:
            $ref: "#/components/schemas/VariableDefinitionDto"
        contextRequirements:
          type: array
          items:
            $ref: "#/components/schemas/ContextRequirementDto"
        extends:
          $ref: "#/components/schemas/ExtendsRefDto"
        validation:
          $ref: "#/components/schemas/ValidationSettingsDto"
        output:
          $ref: "#/components/schemas/OutputDeclarationDto"
      required:
      - contentHash
      - contextRequirements
      - semVer
      - source
      - state
      - validation
      - variables
...
    DependencyEdgeDto:
      type: object
      properties:
        fromKey:
          type: string
        fromVersion:
          type: string
        toKind:
          type: string
        toKey:
          type: string
        toVersion:
          type: string
      required:
      - fromKey
      - fromVersion
      - toKey
      - toKind
...
  securitySchemes:
    ciap-bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT